Initial Commit
Reset this repo and started from scratch so we could have an empty history.

It now primarily provides a setup script to bootstrap new dev machines, and to a lesser extend update existing ones. The repo is public to aid in bootstrapping machines without having to even have Git installed.